test(bmc-mock): move NvidiaDgxVr test to its own module
Per review feedback: NvidiaDgxVr is Vera Rubin hardware, architecturally
distinct from Lenovo GB300, and shouldn't live in
lenovo_gb300_explore.rs. Moved to a new vera_rubin_explore.rs file,
matching the existing hw::vera_rubin naming convention used elsewhere
in the codebase.
